+++ R/sanity.R
@@ -16,7 +16,7 @@ sanity_align <- function(align, i) {
 
 sanitize_i <- function(i, x, pre_group_i = FALSE, lazy = TRUE) {
   if (is.character(i)) {
-    assert_choice(i, c("notes", "caption"))
+    assert_choice(i, c("notes", "caption", "group"))
     return(i)
   } else if (is.matrix(i) && is.logical(i)) {
     return(i)