+++ NEWS.md
@@ -13,6 +13,7 @@ Misc:
 
 * `group_tt()` argument `j` accepts a single string as delimiter (ex: `"_"`) when column names of the data frame include group names. Thanks to @sda030 for the feature request, a first implementation, and testing.
 * `style_tt(i = "groupi")` styles group rows.
+* `theme_revealjs()` now supports dark theme with `css = "dark"` argument.
 * Many improvements to the documentation and vignettes
 * New vignette with table of contents for CRAN
 * `tt()` gets a `colnames` argument, instead of having to delete names manually.